feat: add conflict and not found exception handling in user service methods

+++ backend/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
 
@@ -8,6 +8,11 @@ export class UserService {
     
     async createUser(email: string, password: string) {
         const hashedPassword = await bcrypt.hash(password, 10);
+
+        const existingUser = await this.prisma.user.findUnique({ where: { email } });
+        if (existingUser) {
+          throw new ConflictException('User already exists');
+        }
         return this.prisma.user.create({
           data: { email, password: hashedPassword },
         });
@@ -18,10 +23,18 @@ export class UserService {
     }
 
     async findOne(email: string) {
+        const existingUser = await this.prisma.user.findUnique({ where: { email } });
+        if (!existingUser) {
+          throw new NotFoundException('User not found');
+        }
         return this.prisma.user.findUnique({where : { email : email}})
     }
 
     async update(email: string, data: { email?: string; password?: string }) {
+        const existingUser = await this.prisma.user.findUnique({ where: { email } });
+        if (!existingUser) {
+          throw new NotFoundException('User not found');
+        }
         if (data.password) {
           data.password = await bcrypt.hash(data.password, 10);
         }
@@ -32,6 +45,10 @@ export class UserService {
       }
 
       async delete(email: string) {
+        const existingUser = await this.prisma.user.findUnique({ where: { email } });
+        if (!existingUser) {
+          throw new NotFoundException('User not found');
+        }
         return this.prisma.user.delete({where:{ email:email }})
     }
 }
